refactor: reorganize graphql resolvers and schema
