+++ src/graphql/resolvers/albums/index.ts
@@ -1,0 +1,5 @@
+import { getSavedAlbums } from "./getSavedAlbums.js";
+
+export const albumsResolver = {
+  Query: { getSavedAlbums },
+};
