+++ src/graphql/resolvers/users/index.ts
@@ -1,0 +1,5 @@
+import { getUserProfile } from "./getUserProfile.js";
+
+export const usersResolver = {
+  Query: { getUserProfile },
+};
