+++ src/graphql/schema/common/artist.ts
@@ -1,0 +1,9 @@
+export const ArtistType = `
+  type Artist {
+    external_urls: ExternalUrls
+    href: String
+    id: String
+    name: String
+    type: String
+    uri: String
+  }`;
